refactor(checker): split class index-signature compat into its own module
Move `check_class_index_signature_compatibility` out of
`class_checker_compat.rs` into a dedicated `class_index_signature_compat.rs`
sibling module so the file stays under the 2000-LOC checker-boundary limit
after the polymorphic-`this` override fix. Pure code move; no behavior change.
