+++ crates/tsz-checker/src/classes/mod.rs
@@ -7,6 +7,7 @@ pub(crate) mod class_checker_compat_this;
 pub(crate) mod class_helpers;
 pub(crate) mod class_implements_checker;
 pub(crate) mod class_implements_helpers;
+pub(crate) mod class_index_signature_compat;
 pub mod class_inheritance;
 pub(crate) mod class_member_info;
 pub(crate) mod class_static_side_helpers;
